move em410x demod functions to cmdlfem4x.c
from cmddata.c
NOTE: see changelog.md for command line changes!
